requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.45.0
accelerate>=0.30.0
peft>=0.11.0
bitsandbytes>=0.43.0

# Vision
Pillow>=10.0.0
opencv-python>=4.8.0
qwen-vl-utils>=0.0.8

# Knowledge/RAG
faiss-gpu>=1.7.4
sentence-transformers>=2.7.0

# Data
datasets>=2.19.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Evaluation
evaluate>=0.4.0
rouge-score>=0.1.2
nltk>=3.8.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
loguru>=0.7.0

# Deployment
docker>=7.0.0

# Dev
tensorboard>=2.16.0
